Utilise un concerns sur le modèle Evenement à la place de CreeEvenementAction
